requirements.txt
flask
flask-cors
gunicorn
boto3
python-dotenv
flask-login
email-validator
werkzeug
